third_party/WebKit/LayoutTests/fast/dom/Range/range-exceptions-expected.txt

This test checks some DOM Range exceptions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS node.innerHTML is '<bar>AB<moo>C</moo>DE</bar>'
PASS r.surroundContents(document.createElement('a')) threw exception InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node..
PASS r.surroundContents(document.createElement('a')) threw exception HierarchyRequestError: Failed to execute 'surroundContents' on 'Range': The node to be inserted is a 'A' node, which may not be inserted here..
PASS successfullyParsed is true

TEST COMPLETE
